Show Validation Feedback and Submit Button Behaviour
Due to remote validation, response from the remote source can have delays. As such users are able to submit the form even before the remote validation is made.

This disables the submit behaviour upon validateInput and enables it when runValidators is done. 

Also implemented a gylphicon feedback when fields are in the process of being validated, especially helpful for delays in remote validation. This can be switched on/off in the option be defining the 'validating' key - true by default.